+++ CHANGES.rst
@@ -1,5 +1,6 @@
 1.3 (unreleased)
 ----------------
+* Recenter images if necessary for routine wavefront maintenance. [#93]
 
 1.2 (2021-06-11)
 ----------------